requirements.txt
django==3.0.0
djangorestframework==3.12.2
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.4.0
celery